Rework springdoc modules organization
