+++ src/test/features/springdoc.feature
@@ -2,8 +2,8 @@ Feature: Springdoc modules
 
   Scenario: Should add Springdoc for MVC
     When I apply modules to default project
-      | maven-java        |
-      | spring-boot        |
+      | maven-java            |
+      | spring-boot           |
       | springdoc-mvc-openapi |
     Then I should have "springdoc-openapi-ui" in "pom.xml"
     And I should have files in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc"
@@ -17,8 +17,8 @@ Feature: Springdoc modules
   Scenario: Should add Springdoc for Webflux
     When I apply modules to default project
       | maven-java                |
-      | spring-boot                |
-      | spring-boot-webflux-netty  |
+      | spring-boot               |
+      | spring-boot-webflux-netty |
       | springdoc-webflux-openapi |
     Then I should have "springdoc-openapi-webflux-ui" in "pom.xml"
     And I should have files in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc"
@@ -29,39 +29,20 @@ Feature: Springdoc modules
     And I should have files in "src/test/resources/config"
       | application.properties |
 
-  Scenario: Should add Springdoc for MVC with JWT Security
+  Scenario: Should apply springdoc oauth2 module
     When I apply modules to default project
-      | maven-java                              |
-      | spring-boot                              |
-      | springdoc-mvc-openapi-with-security-jwt |
-    Then I should have "springdoc-openapi-ui" in "pom.xml"
-    And I should have files in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc"
-      | SpringdocConfiguration.java |
-    And I should have "jwt" in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc/SpringdocConfiguration.java"
-    And I should have files in "src/main/resources/config"
-      | application.properties |
-    And I should have files in "src/test/resources/config"
-      | application.properties |
+      | maven-java            |
+      | spring-boot           |
+      | springdoc-mvc-openapi |
+      | springdoc-oauth2      |
+    Then I should have files in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc/"
+      | SpringdocOAuth2Configuration.java |
 
-  Scenario: Should add Springdoc for Webflux with JWT Security
+  Scenario: Should apply springdoc jwt module
     When I apply modules to default project
-      | maven-java                                  |
-      | spring-boot                                  |
-      | spring-boot-webflux-netty                    |
-      | springdoc-webflux-openapi-with-security-jwt |
-    Then I should have "springdoc-openapi-webflux-ui" in "pom.xml"
-    And I should have files in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc"
-      | SpringdocConfiguration.java |
-    And I should have "jwt" in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc/SpringdocConfiguration.java"
-    And I should have files in "src/main/resources/config"
-      | application.properties |
-    And I should have files in "src/test/resources/config"
-      | application.properties |
-
-  Scenario: Should add Springdoc for MVC with OAuth2 Security
-    When I apply modules to default project
-      | maven-java                                 |
-      | spring-boot                                 |
-      | spring-boot-oauth2                          |
-      | springdoc-mvc-openapi-with-security-oauth2 |
-    Then I should have "springdoc-openapi-ui" in "pom.xml"
+      | maven-java            |
+      | spring-boot           |
+      | springdoc-mvc-openapi |
+      | springdoc-jwt         |
+    Then I should have files in "src/main/java/tech/jhipster/chips/technical/infrastructure/primary/springdoc/"
+      | SpringdocJWTConfiguration.java |
