+++ src/test/java/tech/jhipster/lite/generator/server/springboot/apidocumentation/springdocjwt/domain/SpringdocJwtModuleFactoryTest.java
@@ -1,0 +1,29 @@
+package tech.jhipster.lite.generator.server.springboot.apidocumentation.springdocjwt.domain;
+
+import static tech.jhipster.lite.module.infrastructure.secondary.JHipsterModulesAssertions.*;
+
+import org.junit.jupiter.api.Test;
+import tech.jhipster.lite.TestFileUtils;
+import tech.jhipster.lite.UnitTest;
+import tech.jhipster.lite.module.domain.JHipsterModule;
+import tech.jhipster.lite.module.domain.JHipsterModulesFixture;
+import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
+
+@UnitTest
+class SpringdocJwtModuleFactoryTest {
+
+  private static final SpringdocJwtModuleFactory factory = new SpringdocJwtModuleFactory();
+
+  @Test
+  void shouldBuildModule() {
+    JHipsterModuleProperties properties = JHipsterModulesFixture
+      .propertiesBuilder(TestFileUtils.tmpDirForTest())
+      .basePackage("com.jhipster.test")
+      .build();
+
+    JHipsterModule module = factory.buildModule(properties);
+
+    assertThatModule(module)
+      .createFiles("src/main/java/com/jhipster/test/technical/infrastructure/primary/springdoc/SpringdocJWTConfiguration.java");
+  }
+}
